Added receiving administrative commands to korrpcdid.
How commander is provided to Indira is a subject to change, as the current way
is not the best one for various reasons.

The commander doesn't understand any command and always replies with an error.

bin/korrpcdid lacks heavily in error formatting, this is also to be changed.

But hey, it's the first more-or-less functioning daemon with Indira!